+++ modules/src/ics02_client/raw.rs
@@ -1,8 +1,12 @@
-use crate::ics03_connection::error::Kind;
-use crate::ics24_host::identifier::ConnectionId;
 use std::convert::TryFrom;
 use std::str::FromStr;
+
+use serde::Serialize;
+
 use tendermint_proto::Protobuf;
+
+use crate::ics03_connection::error::Kind;
+use crate::ics24_host::identifier::ConnectionId;
 
 //TODO: This might need to be migrated to ibc-proto crate. But ClientConnections (as array of strings)
 // might not be part of an official proto file
@@ -12,7 +16,7 @@ pub struct RawClientConnections {
     pub connections: ::std::vec::Vec<String>,
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Serialize)]
 pub struct ConnectionIds(pub Vec<ConnectionId>);
 
 impl Protobuf<RawClientConnections> for ConnectionIds {}
@@ -21,19 +25,15 @@ impl TryFrom<RawClientConnections> for ConnectionIds {
     type Error = anomaly::Error<Kind>;
 
     fn try_from(value: RawClientConnections) -> Result<Self, Self::Error> {
-        if !value.connections.is_empty() {
-            let mut connections: Vec<ConnectionId> = vec![];
-            for value in value.connections {
-                let conn_id = ConnectionId::from_str(&value.replace("connections/", ""));
-                match conn_id {
-                    Ok(c) => connections.push(c),
-                    Err(_e) => return Err(Kind::IdentifierError.into()),
-                }
+        let mut connections: Vec<ConnectionId> = vec![];
+        for value in value.connections {
+            let conn_id = ConnectionId::from_str(&value.replace("connections/", ""));
+            match conn_id {
+                Ok(c) => connections.push(c),
+                Err(_e) => return Err(Kind::IdentifierError.into()),
             }
-            Ok(ConnectionIds(connections))
-        } else {
-            Err(Kind::ConnectionNotFound.into())
         }
+        Ok(ConnectionIds(connections))
     }
 }