+++ relayer-cli/src/commands/tx/packet.rs
@@ -1,17 +1,19 @@
-use crate::prelude::*;
-
 use abscissa_core::{Command, Options, Runnable};
-use relayer::config::Config;
+use serde_json::json;
 
-use crate::error::{Error, Kind};
 use ibc::events::IBCEvent;
 use ibc::ics24_host::identifier::{ChannelId, ClientId, PortId};
 use relayer::chain::runtime::ChainRuntime;
 use relayer::chain::CosmosSDKChain;
+use relayer::config::Config;
 use relayer::link::{
     build_and_send_ack_packet_messages, build_and_send_recv_packet_messages, PacketEnvelope,
     PacketOptions,
 };
+
+use crate::conclude::Output;
+use crate::error::{Error, Kind};
+use crate::prelude::*;
 
 #[derive(Clone, Command, Debug, Options)]
 pub struct TxRawPacketRecvCmd {
@@ -73,8 +75,7 @@ impl Runnable for TxRawPacketRecvCmd {
 
         let opts = match self.validate_options(&config) {
             Err(err) => {
-                status_err!("invalid options: {}", err);
-                return;
+                return Output::with_error().with_result(json!(err)).exit();
             }
             Ok(result) => result,
         };
@@ -90,12 +91,10 @@ impl Runnable for TxRawPacketRecvCmd {
                 .map_err(|e| Kind::Tx.context(e).into());
 
         match res {
-            Ok(events) => status_info!(
-                "packet recv, result: ",
-                "{:#?}",
-                serde_json::to_string(&events).unwrap()
-            ),
-            Err(e) => status_info!("packet recv failed, error: ", "{}", e),
+            Ok(ev) => Output::with_success().with_result(json!(ev)).exit(),
+            Err(e) => Output::with_error()
+                .with_result(json!(format!("{}", e)))
+                .exit(),
         }
     }
 }
@@ -160,8 +159,7 @@ impl Runnable for TxRawPacketAckCmd {
 
         let opts = match self.validate_options(&config) {
             Err(err) => {
-                status_err!("invalid options: {}", err);
-                return;
+                return Output::with_error().with_result(json!(err)).exit();
             }
             Ok(result) => result,
         };
@@ -177,12 +175,10 @@ impl Runnable for TxRawPacketAckCmd {
                 .map_err(|e| Kind::Tx.context(e).into());
 
         match res {
-            Ok(events) => status_info!(
-                "packet ack, result: ",
-                "{:#?}",
-                serde_json::to_string(&events).unwrap()
-            ),
-            Err(e) => status_info!("packet ack failed, error: ", "{}", e),
+            Ok(ev) => Output::with_success().with_result(json!(ev)).exit(),
+            Err(e) => Output::with_error()
+                .with_result(json!(format!("{}", e)))
+                .exit(),
         }
     }
 }