+++ relayer/src/chain.rs
@@ -23,8 +23,8 @@ use tendermint::account::Id as AccountId;
 use tendermint::block::Height;
 
 use ibc_proto::ibc::core::channel::v1::{
-    PacketState, QueryPacketAcknowledgementsRequest, QueryPacketCommitmentsRequest,
-    QueryUnreceivedAcksRequest, QueryUnreceivedPacketsRequest,
+    PacketState, QueryConnectionChannelsRequest, QueryPacketAcknowledgementsRequest,
+    QueryPacketCommitmentsRequest, QueryUnreceivedAcksRequest, QueryUnreceivedPacketsRequest,
 };
 use ibc_proto::ibc::core::commitment::v1::MerkleProof;
 
@@ -154,7 +154,8 @@ pub trait Chain: Sized {
         height: ICSHeight,
     ) -> Result<ConnectionEnd, Error> {
         let res = self.query(Path::Connections(connection_id.clone()), height, false)?;
-        Ok(ConnectionEnd::decode_vec(&res.value).map_err(|e| Kind::Query.context(e))?)
+        Ok(ConnectionEnd::decode_vec(&res.value)
+            .map_err(|e| Kind::Query("connection".into()).context(e))?)
     }
 
     fn query_channel(
@@ -168,7 +169,8 @@ pub trait Chain: Sized {
             height,
             false,
         )?;
-        Ok(ChannelEnd::decode_vec(&res.value).map_err(|e| Kind::Query.context(e))?)
+        Ok(ChannelEnd::decode_vec(&res.value)
+            .map_err(|e| Kind::Query("channel".into()).context(e))?)
     }
 
     // Provable queries
@@ -186,14 +188,15 @@ pub trait Chain: Sized {
     ) -> Result<(ConnectionEnd, MerkleProof), Error> {
         let res = self
             .query(Path::Connections(connection_id.clone()), height, true)
-            .map_err(|e| Kind::Query.context(e))?;
-        let connection_end =
-            ConnectionEnd::decode_vec(&res.value).map_err(|e| Kind::Query.context(e))?;
+            .map_err(|e| Kind::Query("proven connection".into()).context(e))?;
+        let connection_end = ConnectionEnd::decode_vec(&res.value)
+            .map_err(|e| Kind::Query("proven connection".into()).context(e))?;
 
         Ok((
             connection_end,
-            res.proof
-                .ok_or_else(|| Kind::Query.context("empty proof".to_string()))?,
+            res.proof.ok_or_else(|| {
+                Kind::Query("proven connection".into()).context("empty proof".to_string())
+            })?,
         ))
     }
 
@@ -216,14 +219,16 @@ pub trait Chain: Sized {
                 height,
                 true,
             )
-            .map_err(|e| Kind::Query.context(e))?;
+            .map_err(|e| Kind::Query("proven channel".into()).context(e))?;
 
-        let channel_end = ChannelEnd::decode_vec(&res.value).map_err(|e| Kind::Query.context(e))?;
+        let channel_end = ChannelEnd::decode_vec(&res.value)
+            .map_err(|e| Kind::Query("proven channel".into()).context(e))?;
 
         Ok((
             channel_end,
-            res.proof
-                .ok_or_else(|| Kind::Query.context("empty proof".to_string()))?,
+            res.proof.ok_or_else(|| {
+                Kind::Query("proven channel".into()).context("empty proof".to_string())
+            })?,
         ))
     }
 
@@ -353,6 +358,12 @@ pub trait Chain: Sized {
         request: QueryUnreceivedAcksRequest,
     ) -> Result<Vec<u64>, Error>;
 
+    /// Performs a query to retrieve the identifiers of all channels associated with a connection.
+    fn query_connection_channels(
+        &self,
+        request: QueryConnectionChannelsRequest,
+    ) -> Result<Vec<ChannelId>, Error>;
+
     fn build_packet_proofs(
         &self,
         packet_type: PacketMsgType,
@@ -388,13 +399,12 @@ pub trait Chain: Sized {
 
         let res = self
             .query(data, height, true)
-            .map_err(|e| Kind::Query.context(e))?;
+            .map_err(|e| Kind::Query(packet_type.to_string()).context(e))?;
 
         let proofs = Proofs::new(
-            CommitmentProofBytes::from(
-                res.proof
-                    .ok_or_else(|| Kind::Query.context("empty proof".to_string()))?,
-            ),
+            CommitmentProofBytes::from(res.proof.ok_or_else(|| {
+                Kind::Query(packet_type.to_string()).context("empty proof".to_string())
+            })?),
             None,
             None,
             height.increment(),