+++ CHANGELOG.md
@@ -2,20 +2,28 @@
 
 ## Unreleased Changes
 
-- Replace `ChannelConfig` in `Channel::new` ([#511])
 
 ### FEATURES
 
 - Add support for streamlining releases ([#507])
+- [relayer-cli]
+  - Implement command to query the channels associated with a connection ([#505])
+  - JSON output for queries and txs ([#500])
 
 ### IMPROVEMENTS
+
+- [relayer-cli]
+  - Replace `ChannelConfig` in `Channel::new` ([#511])
 
 - Update to `tendermint-rs` v0.17.1 ([#517])
 
 
+[#500]: https://github.com/informalsystems/ibc-rs/issues/500
+[#505]: https://github.com/informalsystems/ibc-rs/issues/505
 [#507]: https://github.com/informalsystems/ibc-rs/issues/507
 [#511]: https://github.com/informalsystems/ibc-rs/pull/511
 [#517]: https://github.com/informalsystems/ibc-rs/issues/517
+
 
 ## v0.0.6
 *December 23, 2020*